refactor: give Tensor its own file
